requirements.txt
pandas
psycopg2
openpyxl
xlsxwriter
requests
python-dotenv
